+++ plugin_info/install.php
@@ -42,7 +42,23 @@ require_once dirname(__FILE__) . '/../../../core/php/core.inc.php';
             }
             if (empty($eqLogic->getConfiguration('counter'))) {
                 $eqLogic->setConfiguration('counter', 'auto');
-            }     
+            }
+            for ($i = 1; $i <= 12; $i++) {
+                if ($eqLogic->getConfiguration('month_'.$i, 'unset') === 'unset') {
+                    $eqLogic->setConfiguration('month_'.$i, 1);
+                }
+            }
+            if ($eqLogic->getConfiguration('color') === 'braun') {
+                $eqLogic->setConfiguration('color', 'brown');
+            }
+            if ($eqLogic->getConfiguration('notif_veille') == 1) {
+                $eqLogic->setConfiguration('notif_days', 1);
+            } 
+            if ($eqLogic->getConfiguration('notif_veille') == 0) {
+                $eqLogic->setConfiguration('notif_days', 0);
+            }
+            $eqLogic->setConfiguration('notif_veille', 'unused');
+            
             $cmd = $eqLogic->getCmd(null, 'counter');
             if (!is_object($cmd)) {
                 $cmd = new mybinCmd();