+++ docs/en_US/changelog.md
@@ -4,6 +4,11 @@
 >
 >As a reminder if there is no information on the update, it means that it only concerns the updating of documentation, translation or text.
 
+# 12/03/2021
+- Possibility to specify months
+- Possibility to define the number of days before collect for notification
+- Fix on brown color
+
 # 08/03/2021
 - New bin color: black
 - New icon color for no bin (for better visibility on white background)